Fix type hints

+++ vllm/model_executor/layers/attention/attention.py
@@ -54,7 +54,7 @@ class Attention(nn.Module):
         key_cache: Optional[torch.Tensor],
         value_cache: Optional[torch.Tensor],
         input_metadata: InputMetadata,
-        kv_scale: int = 1.0,
+        kv_scale: float = 1.0,
     ) -> torch.Tensor:
         return self.backend.forward(query, key, value, key_cache, value_cache,
                                     input_metadata, kv_scale)
