+++ vllm/model_executor/layers/attention/backends/flash_attn.py
@@ -48,7 +48,7 @@ class FlashAttentionBackend:
         key_cache: Optional[torch.Tensor],
         value_cache: Optional[torch.Tensor],
         input_metadata: InputMetadata,
-        kv_scale: int,
+        kv_scale: float,
     ) -> torch.Tensor:
         """Forward pass with FlashAttention and PagedAttention.
